float64 in the tens calculation
saw imprecise results when using float32

+++ elePyant/core.py
@@ -46,7 +46,7 @@ def _round_nsd_base10(x, nsd):
     """
     where_zero = x == 0  # avoid taking log10 of 0
     tens = np.ones_like(x)
-    tens[~where_zero] = 10 ** np.ceil(np.log10(np.abs(x[~where_zero])))
+    tens[~where_zero] = 10 ** np.ceil(np.log10(np.abs(x[~where_zero].astype(np.float64))))
     # note: could instead pass `where=where_zero` to np.abs or np.log etc.
     
     # "true normalized" significand
